tools/requirements.txt
anthropic>=0.40.0
jinja2>=3.1.0
pyyaml>=6.0
python-slugify>=8.0.0
pillow>=10.0.0
